industrializing evolutionary evaluations

+++ NCEP_si_verf/year_cice.py
@@ -28,14 +28,17 @@ for evo in range(0,nexpt):
   EXPT='/gaea_intel_smoke_gx3_1x1_evo'+sno+'_med3_yr_out.'+GEN
   fname = 'fout'+sno
 
-  fout = open(fname,'w')
-  while (tag <= end):
-    dtag=tag.strftime("%Y-%m-%d")
-    cmd = EXDIR+'/cice_solo '+FIXDIR+'/seaice_gland5min '+OUTDIR+EXPT+ '/history/iceh.'+dtag+'.nc '+"{:.3f}".format(acrit) + ' ' + dtag + " >> "+ fname
-    #debug: print(tag," ",end="", file = fout, flush=True)
-    #debug: print(tag,flush=True)
-    #debug: print(cmd, flush=True)
-    os.system(cmd)
-    tag += dt
-  fout.close()
-  #debug: exit(0)
+  if (not os.path.exists(fname)):
+    fout = open(fname,'w')
+    while (tag <= end):
+      dtag=tag.strftime("%Y-%m-%d")
+      cmd = EXDIR+'/cice_solo '+FIXDIR+'/seaice_gland5min '+OUTDIR+EXPT+ '/history/iceh.'+dtag+'.nc '+"{:.3f}".format(acrit) + ' ' + dtag + " >> "+ fname
+      #debug: print(tag," ",end="", file = fout, flush=True)
+      #debug: print(tag,flush=True)
+      #debug: print(cmd, flush=True)
+      os.system(cmd)
+      tag += dt
+
+    fout.close()
+    #debug: exit(0)
+
